+++ utils/fixed_decimal/src/lib.rs
@@ -51,19 +51,23 @@
     )
 )]
 
+mod compact;
 mod decimal;
+mod integer;
 mod ops;
+mod scientific;
 mod uint_iterator;
 
 #[cfg(feature = "ryu")]
 pub use decimal::DoublePrecision;
 
-pub use decimal::CompactDecimal;
+pub use compact::CompactDecimal;
 pub use decimal::FixedDecimal;
-pub use decimal::ScientificDecimal;
 pub use decimal::Sign;
 pub use decimal::SignDisplay;
 use displaydoc::Display;
+pub use integer::FixedInteger;
+pub use scientific::ScientificDecimal;
 
 #[derive(Display, Debug, PartialEq)]
 #[non_exhaustive]
@@ -71,6 +75,9 @@ pub enum Error {
     /// The magnitude or number of digits exceeds the limit of the FixedDecimal. The highest
     /// magnitude of the most significant digit is core::i16::MAX, and the lowest magnitude of the
     /// least significant digit is core::i16::MIN.
+    ///
+    /// This error is also returned when constructing a FixedInteger from a FixedDecimal with a
+    /// fractional part.
     ///
     /// # Examples
     ///
@@ -93,9 +100,6 @@ pub enum Error {
     /// 123 (or 123.0) must be used.
     #[displaydoc("Failed to parse the input string")]
     Syntax,
-    /// An exponent with fractional digits is passed when constructing a [`ScientificDecimal`].
-    #[displaydoc("Exponent must not have fractional digits")]
-    IntegerExponent,
 }
 
 #[cfg(feature = "std")]
